moved KeyCon and miscIcons folders into jsicons (from js icons)
